requirements.txt
aiohttp==3.11.13
bs4==0.0.2
numpy==2.2.3
pandas==2.2.3
pymongo==4.11.1
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.38
waitress==3.0.2
Flask==3.1.0